Kludge: dbms.threads.worker_count
The setting dbms.threads.worker_count is set with an expression,
evaluated as the setting is defined.
The correct thing to document as the default value is "The minimum between
'number of processors' and 500".
Instead the value is computed on the build server and comes out
(usually) as "2".

This is a glorious kludge to replace the default value for this setting when
generating config docs.