Fix broken JMX link
The index file including generated content was out of synch with the
generated content. Now generates the includes as well to keep them in
sync in future.

Also rewrite the test to make it a little easier to understand and work
with. (This really should be a tool with a CLI, not a unit test.)