Update terminology for lookup index -> token lookup index
This is to better align with the user surface which uses "token index"
when describing those indexes. "Lookup" only describe what type of
predicates the index has support for.

It's also to protect us from a future scenario where we might want
to have a "lookup" index on properties. That would then be a "property
lookup index" or similar.

Also updates to use b-tree consistently